requirements.txt
PyYAML>=6.0,<7.0
pytest>=8.0,<9.0
pandas>=2.0,<3.0
requests>=2.31,<3.0
yfinance>=0.2,<1.0
lxml>=5.0,<6.0
pyarrow>=14.0,<20.0
